Ensure using stable source IP to connect to OpenShift bootstrap node
In an OpenShift UPI dual ToR deployment, every node is dual-homed -
and so has a stable address provisioned for it - except for the
bootstrap node.

The bootstrap node cannot be dual-homed because there is no way to
control the IP that is exported to other nodes for the bootstrap API
server.  (At least, this was the case as of OpenShift 4.5, and we
haven't re-investigated for OpenShift 4.6 and 4.7.)  Also in the
bigger picture there is no need for the bootstrap node to be
dual-homed, as it is deleted once the real control nodes are up.  The
important thing is that all the persistent cluster nodes are
dual-homed.

During the bootstrapping phase, the real control nodes need to access
the bootstrap etcd server, and there can(*) be a problem with access
from a control node in the same rack as the bootstrap node because the
bootstrap etcd server sees the control node connecting to it with the
wrong source IP.  The bootstrap etcd server expects the control node's
source IP to be its loopback IP (e.g. 172.31.20.2) but what happens
is:

- The control node wants to route to the bootstrap IP (e.g. 172.31.21.1).

- The control node is also on the 172.31.21/24 network, so it routes
  directly on that interface and chooses its own IP on that network
  as the source, e.g. 172.31.21.2.

- Boostrap etcd server rejects the attempted connection like this:
  ```
  I | embed: rejected connection from "172.31.21.2:48460" (error "tls: \"172.31.21.2\" does not match any of DNSNames [\"localhost\" \"172.31.20.2\"]", ServerName "", IPAddresses ["172.31.20.2"], DNSNames ["localhost" "172.31.20.2"])
  ```

(*) For some reason this does not happen _every_ time that I try to
bring up an OpenShift 4.7 rig, but I would estimate it happens on
about 60% of attempts.

To avoid that problem we can add a more specific route to the
bootstrap IP with a source route hint, on each node that is on the
same subnet as the bootstrap.  For example, if we were doing
this on the Linux command line:

```
ip route add 172.31.21.1/32 dev enp0s3 src 172.31.20.2
```

This PR adds config and coding to do that.