Separate Dex env variable from OIDC related env variables
We now have a separate auth flow from dex for implicit OIDC, so it doesn't make sense to say everything is under Dex. We only need to know if Dex is enabled so we can add the trusted cert for Dex. Just to note as well, all our components use OIDC to authenticate with dex, so it's not really dex authentication, it's OIDC authentication (to dex).