[SAAS-1459] Regenerate managed cluster user secrets if the UID for the ManagedCluster has changed
This PR ensures that the Elasticsearch credentials are recreated and inserted into Elasticsearch if a ManagedCluster resource is deleted then reconnected. When a Managed Cluster resource is deleted, the users are removed from Elasticserach but the user secrets are not removed from the Managed Cluster, meaning that if the Managed Cluster is reconnected it would recreate the Elasticsearch credentials.