Tighten the handling for an egress gateway return flow
When egress gateway and client are on different nodes, we need to
allow a mid-flow TCP packet on the return path to create CT state for
the original unencapped flow, and whitelist that CT state's FROM_HOST
side.

However, the previous code was too broad, potentially also allowing
this for a TCP SYN on the outbound path, which could lead to skipping
HEP policy on the egress node.

Here we make sure that the allowance only applies for mid-flow TCP.