collector UT: Fix health checking flake
Specifically this:

    • Failure [15.001 seconds]
    Flowlog Reporter health verification Test with client that times out requests [It] verify health reporting.
    /go/src/github.com/projectcalico/felix/collector/reporter_flowlogs_test.go:398
      Timed out after 15.000s.
      Expected
          <health.HealthReport>: {Live: true, Ready: true}
      to equal
          <health.HealthReport>: {Live: true, Ready: false}
      /go/src/github.com/projectcalico/felix/collector/reporter_flowlogs_test.go:401

The flow logs Reporter is started with a map of configurable
dispatchers, and the point here is that the Reporter should report
non-ready if any of its dispatchers fail to initialize.  This comes
from when we had a dispatcher - for CloudWatch - that could sometimes
fail to initialize, because of a misconfigured CloudWatch API client.

Recently in #1224 the CloudWatch dispatcher was removed, and by
accident we were left with two identical tests: one expects
'eventually' to see a True readiness report, and the other expects
'eventually' to see a False readiness report.  In practice, both of
these can often pass, because of the calculated readiness transiently
passing through both True and False.

However, what we really want is to reinstate the test for a dispatcher
that fails to initialize, and that's what this PR does.