tproxy: pod-service-self fix
We maintain an ipset hash:net,net with tuples of local pod's IPs. Both
IPs in the tuple are the same. Thus we can compare whether a packet is a
loopback, that is, the pod is trying to connect to self.

Normally, when a pod is trying to connect to self using a service,
kube-proxy sets a MASQ mark (set by --iptables-masquerade-bit) as
otherwise connectivity from the pod-service would not work as (a) rpf &
accept_local usually do not permit receiving packets from outside the
pod to self and (b) return traffic would  not exit the pod and thus
connectivity would be broken.

When we see traffic with a non-local source address in OUTPUT chain on
the host, we assume that it is proxied traffic. If such a traffic has
the same source and destination, we set the kube-proxy's MASQ mark so
that POSTROUTING chain MASQs the traffic in the same vein as if we
preserved the mark across the proxy set in nat PREROUTING - which is not
technically possible.