tproxy: NodePorts
Nodeport traffic is steered to a different port on the proxy so that the
proxy can mark that traffic with the kube-proxy masq bit. This way we
preserve the bit that would be set by kube-proxy itself on the
downstream connection, but is lost on the upstream.

The mark on the proxy must be in sync with kube-proxy.