tproxy: EKS nodeport fix
Because the calico rules preced (by default) any other rules, TPROXY is
executed before AWS CNI rules can mark the nodeport traffic.

This is a hardcoded fix that mimics the EKS rule to mark any connection
that comes from eth0 (the main iface) to a local address (nodeport) with
0x80 (the default EKS value).

We restore the the mark from conntrack in OUTPUT as it is restored if it
comes from eni+ ifaces.

It has no side effect unless it is proxied traffic and thus the source
is pod's source.