Suppress DNS-LATENCY log spam
Despite previous attempts to improve the situation, customers are
still reporting copious amounts of DNS-LATENCY log spam.
E.g. https://tigera.atlassian.net/browse/CI-786.

Two reasons for this are:

1. We don't currently get a timestamp on DNS requests sent from a
host (as opposed to a pod).  In the latency-processing code, the
timestamp in such cases appears as 0.

2. Commit aa5c3b8f59 (by me) probably made things worse by entering
such 0 timestamps into the `s.requestTime` map - which is problematic
we have no way of knowing when to remove those entries again from the
map.

(In detail, the kind of thing that goes wrong is:

- We see a DNS request with no timestamp, add entry into the map with
  0 timestamp.

- We see another DNS request with a good timestamp, add it in too.

- That second request triggers the loop that is trying to delete old
  map entries.  This finds the first request (that in fact only just
  happened), removes it, and logs "Missed DNS response for request"
  warning.

- We see the DNS response for the first request, but now there's no
  matching map entry, so we log "Missed DNS request for response"
  warning.)

Fundamentally we can't make this code really right until we have a way
of getting a comparable timestamp on _every_ DNS packet that we
snoop.  https://tigera.atlassian.net/browse/BPF-730 tracks doing
that.  In the meantime:

- We revert adding zero timestamps into the `s.requestTime` map.

- We downgrade the "Missing timestamp" and "Missed DNS request for
  response" warning logs to debug, because they are expected for a DNS
  lookup from the host.