tproxy: Fix host networked backends - ClusterIP
When a services is backed by a host networked pod and the client happens
to be on the same node, the upstream connection does not go through
PREROUTING but OUTPUT. OUTPUT needs to reroute the response packet to
the proxy as PREROUTING does. For that we need to know that a connection
is upstream so that we do not reroute downstream packet in OUTPUT as
well.

Envoy marks the upstream connection so that we can tell them apart.