Restart nfqueue connection if it closed
It's possible that the nfqueue library we use closes the nfqueue connection if there's an unrecoverable error, and it doesn't re open it. This PR adds logic to detect when that happens, and recreates the nfqueue connection along with the dns policy packet processor.

This commit also adds a debug console attached to a socket. You can use netcat to send commands to the debug console when it's enabled. The only command right now is to close the nfqueue connection, so that we can test that the nfqueue connection restart logic is working.

Here is an example of running this commamd: `echo "close-nfqueue-conn" | nc -U /var/run/calico/felix/debug`