Ensure that the trusted DNS servers are used for the DNS lookups in the fv tests
It's possible that there's a mismatch between the default nameserver that golang calculates (in dns-test.go) and what we tell felix to trust in our test. When this happens, we'll not capture the DNS response in our policy tests which results in the failures we've seen.

This PR changes forces the test-dns.go script to use the dns server that we've told felix to trust.