Fix potential race in DNS policy test (restarting the packet processor)
There's a possible race when we test that the packet processor recovers from an error where we might send a packet too soon, so it will reach the packet processor as it's shutting down and while it's shut down (the later means there's no listener for nfqueue so the packet is auto dropped).

This commit adds a 1 second sleep to give the packet processor time to shutdown and restart. If it takes longer that 1 second then something is likely wrong.