Add post SNAT port(s) to flow logs
This commit adds the post_snat_ports field to the flow logs. This field contains a, possible truncated, list of post SNAT'd ports for the connections in a flow log.

This is being added so that systems outside of the IPPool of an SNAT'd package can correlate seen SNAT'd packets to flow logs (for instance firewalls).

By default a maximum of 3 post SNAT ports will be logged for a flow containing multiple connections, but this can be changed with the FlowLogsFilePostSNATPortLimit felix configuration parameter.

Overwrite flow refs on new metric updates (even if it already exists) to get updates nat outgoing port

Generate CRDs when calling image-test target