BPF mode: Handle policy update for policy in non-"default" tier
When we stored a map from policy IDs to relevant interfaces, we
hardcoded a tier name of “default”.  But when we later looked up in
that same map (`onPolicyUpdate`) we used the correct tier name.  So
for tiers other than "default", the lookup didn't match the store, and
the effect was failing to implement a policy update in the BPF
dataplane.

In fact we don't need any tier name in the map key, because the policy
"Name" field is the full policy name, including the tier name.