Read back AWS state after update.
The AWS API is "eventually consistent" meaning that it can
return stale data.  Occasionally, it also seems to lose "adds"
(possibly when they conflict with recent "deletions").  Handle
both cases by checking that the state looks in sync after an
update and then triggering a backoff/retry if not.

Tune some logs that were helpful in diagnosing the issue.