Remove "Invocation" field from report
Although the Invocation field has always been part of the report spec,
we have never actually implemented it yet, and it is not entirely
straightforward to do so.  I had a go yesterday, and the issues are:

1. The Invocation string is derived from the ProcessStart event and
then cached so that we can also fill it in on other kinds of event.
But when then happens if ProcessStart events are filtered out?

2. AnyArgsForSameCommand aggregation is achievable but a bit fiddly.
For that aggregation to work correctly, the collector needs to trim
Invocation down to just the first arg, and it can't reliably use "look
for the first space" to do that.  So we need the generator to pass
along two fields, "Invocation" and "InvocationWithoutArgs".  That
works, but doesn't feel entirely nice.

Since the v1 use case is based only on ProcessStart reports, and those
already have equivalent (or superior) information to "Invocation", I
think the clearest thing to do is to remove Invocation for now, so as
not to leave a false impression for the other report types.