felix: create DomainInfoStore.MsgChannel at creation time
When BPF events poller gets a packet and tries to send it to
DomainInfoStore before it is properly initialized, we get a send to a
nil channel, which blocks forever.

DomainInfoStore isn't initiated fully until datapalnes Start() executes.
However bunch of things are already started in NewIntDataplaneDriver()
and perhaps should not be or all of them should be. In any case, it does
not hurt to initialize fully at creation time.