Remove description fields from prometheus CRDs
This is a workaround to the following upstream issues:
https://github.com/kubernetes/kubernetes/issues/82292
and
https://github.com/prometheus-operator/prometheus-operator/issues/4439

The command used with yq (4.16.2)

`for f in `ls -1`; do yq -i eval 'del(..| select(has("description")).description)' $f; done`

in the appropriate CRD folder after removing some of the helm templating
lines. These lines have to be restored after as well.