FV: nat outgoing ports for an SNAT'd flow: allow connectivity retries
This test case was disallowing the normal retries for connectivity checking, and in combination with
Felix now taking a bit longer to restart than it used to, this meant the allowed connectivity checks
were happening before the dataplane was programmed, and so the test failed.

Fix is to allow connectivity retries for the usual period (10s).