Fix tesla build when no tenant ID is provided
Contrary to the comment in kube-controllers/cmd/kube-controllers/cloud.go stating that the tenant ID is optional for cloud / tesla, it is currently required and the roles / users are modified to include the tenant ID. This means that currently the tesla build cannot be used for new calico cloud only features.

This commit makes the tenant ID truly optional for the tesla build, and creates the roles and es users as they would be in ee if the tenant ID is not provided.