Throttle runaway for loop in kube controllers worker
The for loop that waits until the controller from the index informer
didn't have a sleep in it, meaning that if the controller couldn't sync
then the CPU usage would go through the roof.

A common case we're runing into where the worker can't sync because it
can't list resources it's supposed to watch.

We use cache.WaitForNamedCacheSync to wait for the caches to sync, it
does an initial poll and waits 100ms, but will also terminate if the
stop channel is closed.