Fix preemptive cluster role deletion in rbac cache
We are deleting the cluster role from the internal cache of the rbac
cache when the cluster role is deleted, however, it should still remain
because cluster role bindings exist that reference it (a cluster role
binding can reference a cluster role that doesn't exist).

This commit marks the cluster role as not existing and clears out the
rules while keeping the subjects in the rule in tact. When the last
cluster role binding the references the "deleted" cluster role is
removed then the cached entry is removed.