Update events index mapping and add index param
This change adds a `MaybeUpdateIndexMapping()` in LMA to update index
mappings when lifecycle isn't enabled. This applies to events index only
at the moment. For other indices with lifecycle enabled, we leave old
data as they were. In order to dismiss or delete event for both old (<
Calico Enterprise v3.12) and new (>= Calico Enterprise v3.12) events
indices, we also need to explicitly set the `index` parameter.