Move ids controller Dockerfile into ids folder
`ARCHES` aren't correctly detected in `lib.Makefile` where the
Dockerfile is one level deeper in `docker-image` folder. This
causes `VALIDARCHES` to be empty and `push-manifests` to fail
with missing `--platforms` errors.

Moving the IDS controller Dockerfile to the top-level ids folder helps to
detect `ARCHES` correctly. This change is similar to kube-controllers
which contains both top-level Dockerfiles and `docker-images` folder for
other necessary Dockerfiles.

This change also fixes `RELEASE_REGISTRIES` for TESLA builds.