Add operatorNamespace to ManagedClusterSpec
The ManagedCluster resource needs to have the managed cluster's operator
namespace so it can be used by the apiserver to generate the
tigera-managed-cluster-connection secret in the correct namespace.

Currently that secret is created in tigera-operator which does not work
for all clusters. Some clusters run the tigera-operator in a different
namespace.