Include refresh=wait_for in elastic bulk request
The original backend bulk processor implementation didn't include a wait
but returned necessary details regarding the dismissed security events
immediately back to the UI. For Elastic near realtime search model [1],
it is possible for an imediate search after update to return outdated
data.

Due to how pagination is implemented between the manager UI and
backend, when a bulk event dismiss request is sent from UI, it needs a
refresh (query backend again) to correct the pagination. This causes an
issue where a dismissed security event remains in the Alert page grid
until next page reload.

The workaround in this changeset is to re-implement the bulk operations
using `BulkSerivce` directly and includes the `refresh` parameter in bulk
requests send to Elastic. It waits for the changes made by the request to be
made visible by a refresh before replying to UI. The `BulkProcessorService`
still doesn't support this parameter as of today so we are not using methods
from LMA.

[1] https://www.elastic.co/guide/en/elasticsearch/reference/current/near-real-time.html#near-real-time