Buffer zip archive and ignore errors when creating them
We noticed zip archive corruption when it is downloaded from Tigera
manager UI, especially when multiple pcap files are in one archive. The
corrupted zip archive is none-zero sized and contains a few good pcap
files within it when it is analysed using some hex editors.

Here is the current behavior for tigera-packetcapture server to send a
zip archive when it gets the download pcap request from UI:

1. It will first find an entry pod for the pcap files.
2. It runs a remote `tar` command and copy data back to a zip writer.
3. The zip writer will write back to http response writer as soon as the
   first valid pcap file is received.

From the user browser, it will get a 200 OK response when the first bit
is received from the zip writer. However, this `GET` request is still in
a pending state since there are more files to compress and stream to the
client. If any of the above steps failed, the tigera-packetcapture
server sets the error message and close the http connection. This leaves
a partially compressed and corrupted zip archive to the user.

This changeset changes the zip archiving behavior to save the
compressed stream to a buffer first and ignores errors from any of
the above steps. The server will compress and send pcap files as many as
possible to the client. The compress method for zip is also correctly
set to `zip.Deflate`.

Client browser will see a pending state when the `GET` request is made.
It will download a complete zip archive only after it is created successfully.
If any errors occurred during the zip creation, an internal server error will
be returned.