Use MD5 to calculate managed cluster cert fingerprint
This changeset fixes an issue in managed cluster certificate fingerprint
calculation when FIPS mode is enabled. In the initial work on FIPS mode
in [1], we modify the `GenerateFingerprint` helper function in voltron
to use SHA256 when FIPS mode is enabled. However, when the
ManagedCluster resource is created, we always use MD5 to calculate the
managed cluster certificate fingerprint (in libcalico-go
`managedCluster_storage.go`). Using SHA256 when FIPS mode is enabled
causes server side validation error and rejects the client side
connection.

This changeset also cleans up Dockerfiles for both images.

[1] Initial FIPS mode work: https://github.com/tigera/calico-private/pull/4895