Use SHA256 for ManagedCluster active fingerprint
This changeset upgrades the ManagedCluster fingerprint hash algorithm
from md5 to sha256. In Calico Enterprise v3.15, we introduced FIPS mode.
md5 hash algorithm is not approved when FIPS is enabled and it is known
to cause collisions. After this change, sha256 will be used to hash
ManagerdCluster certificate fingerprints no matter what FIPS mode is
used. When a voltron/guardian tunnel is re-established, if voltron sees
an old md5 fingerprint, it will validate the old md5 sum first and update
that to use sha256 hash algorithm.