Change /usr/share/elasticsearch owner to elasticsearch
This changeset modifies `/usr/share/elasticsearch` folder permission
from `root:root` to `elasticsearch:elasticsearch`. The `chown` command
using UID and GID `0` is from the upsteam dockerfile. It works on
non-SELinux enabled hosts like Ubuntu but not on RHEL hosts which is
enabled by default.

When our new UBI based Elasticsearch image is deployed on Openshift,
we got errors on `java` not found. This is actually a permission error
as `/usr/share/elasticsearch/jdk/bin` is set to 0750 which blocks
elasticsearch user from accessing files under the `jdk` folder.