Check file existence before recreating the keystore
This fixes an null input exception in jdk keytool when nodes are
powered off during night and powered on again in the next morning.
keytool exception causes `elastic-internal-init-keystore` container to
crash and `elasticsearch` container to be not ready. Do note that the
default k8s liveness will kick in after a certain timeout and this issue
will be resolved automatically.