Move kube-bench into calico-private and adjust timeouts
This changeset moves kube-bench from a separate tigera customised base
image into calico private repository to streamline compliance benchmarker
container build. The main entrypoint from kube-bench is unlikely to
change very often. We only need to update the kube-bench version in the
top-level `go.mod` for future upgrades. It is also in Apache license,
to my best knowledge, we won't have much legal obligations to include it
in our own repository.

The reporter timeout is also adjusted (increased) as the processing time
is much longer than the default in a large/busy cluster. It is very likely
to be not live and terminated when processing large kubernetes logs.