Use official alpine image in docs and tests
`laurenceman/alpine` alpine image is created 4 yrs ago by Laurence who
left company already. We'd better not reference custom or priviate
container images in our official doc or tests. The official `alpine`
image is used instead.

The custom alpine image is at [1] and the changes in that image is
adding `curl` and a super long sleep entrypoint.

* For egress documentation, the `nc` (netcat) command in Laurence's alpine
image is from busybox so we are safe to switch.
* For node dns policy tests, it uses `wget` which is installed separately.

[1] https://hub.docker.com/layers/laurenceman/alpine/latest/images/sha256-248b0c14fa58314223bafc7ff3dfdd7118c5d2d7db761c98961201a82ddc9fcd?context=explore