Add linseed component (#5430)
* Add linseed component

The new Linseed component will be a REST API exposing endpoints to work with the key resources stored in ElasticSearch. Currently, all clients of es-gateway speak raw ES queries which are proxied by es-gateway to elasticsearch. As part of this design, each of those requests will be replaced with a proper API endpoint that includes a schema, validation, authentication, and authorization.

* [Code Review] Update docs, Makefile and Dockerfile

* [CODE REVIEW] Use secure client and refactor tests for multiple inputs

* Update sem files for ubuntu 2004