[EV-3005] Cannot initiate a packet capture when using Dex for authentication (#5532)
Scratch removes all shared lib64 libraries from the image that will be
used at runtime. When trying to resolve dex.svc.local, packet capture
will need to make calls to resolve this service and localhost.
PacketCapture will be compiled to be statically linked, so we do not
depend on nss shared libraries.