Return empty endpoint list when policy isn't found
This changeset modifies query server endpoints search response.
When a Manager user specifies a valid policy name in Endpoint Details
page but it can't be found in the cluster, it is more common to return
an empty list of items (200 OK) rather than telling users that their
requests are bad (400 Bad Request).