Fix failed endpoint count in query server cache
This changeset adds a Pod deletion callback to QueryServer. When a WEP
is in the Failed phase and removed from the cluster, we add the WEP into
the failedEndpoints collection and report the count back to Manager. For
failed Pods, the API objects still remain in the cluster's API. As the
WEP is removed already, QueryServer endpoints cache won't get updates
when a human or controller process explicitly removes them. A separate
notification function is added to listen to Pod deletion events.