[IA-477] adding changes to RBAC flow for Image Assurance API tokens (#5749)
Starting from v1.24 kubernetes will no longer auto generate API token (item 3) for service accounts, we have to explicitly create a token for each service account that is not mounted on pods.

This PR adds changes so that the current RBAC flow for managed cluster components is not interrupted. With this flow we create SA tokens explicitly for all the IA components with the same name as secret account and populate it in their respective namespace. When we access this secret in operator instead of relying on sa.Secrets[0], we access the secret directly with the SA name..