Upgrade to Kibana 7.17.9 and use official submodule
This changeset upgrade Kibana to v7.17.9 release. The Kibana submodule
is now tracking upstream release branch as we are very inconsistenly
patching and maintaining our fork [1] over the years. Tigera
customizations for Kibana is now created as a patch and applied before
building our own Kibana image. For future updates to the patch file,
we can reapply this patch to upstream Kibana target release, modify it,
and update it back to this repository.

[1] https://github.com/tigera/kibana/tree/tigera-7.17