Regenerate IA Service Account tokens before they expire
Previously we had thought that we could set the Service Account token expire to 100 years, effectively making it unexpiring. We have recently found out that the issuer service in kubernetes doesn't allow service account tokens to have an expiry greater than 48 hours by default.

To temporarily get around this short expiry, we regenerate the tokens before they expire and propagate them to the components that need them.

In the future, we will not use service account tokens for the IA API, as it doesn't seem like they are suitable anymore.