make sure that policy programs are released when iface down.
Policy programs are held in pinned jump map so they do not get released
automatically. We must make sure that when an interface goes down, we
also clean up the entries in the jump map and thus release the programs.

The programs would get released after the entries were overwritten.