Ignore proto in L3 selector for SG service and service group node
This changeset ignores the `proto` field in L3 selector for Service
Graph service (`svc` type) and service group (`svcgp` type) nodes. In
Service Group, flow logs contribute to edges in UI. It can be either
a TCP flow or a UDP flow or both. When generating a L3 selector for a
svc/svcgp node, `proto` is `AND`'ed into the selector expression when
available. This causes discrepancy when the k8s service protocol is TCP
but there are denied UDP traffic (denied DNS requests for example)
associated with it. A user will see the edge in the Service Graph canvas
but no flow logs in the bottom table.