Don't risk truncating DNS packets when copying from iptables to Felix
We've seen DNS policy failing to work - i.e. an outbound connection is not
allowed - when the DNS response is larger than (around) 1024 bytes, and it's
because we were only copying up to 1024 bytes of a DNS response packet from
iptables (NFLOG) to Felix.  There was never any reason for this limit; I think
that when I first implemented this code I was following the pattern of the
NFLOG coding for flow logs - which specifies a packet size of 80 - and I
thought I needed to specify _some_ packet size, and that 1024 would always be
enough.

But it isn't.  The EDNS0 extension allows UDP DNS payloads larger than 512, and
we've now seen this in practice in customer clusters.

There is another required ingredient for this problem: the Linux kernel must be
>= 4.8.  If the Linux kernel is < 4.8, Felix uses the --nflog-range option,
which doesn't actually do anything.  For kernel >= 4.8, Felix uses --nflog-size
instead, and that _does_ limit the size of the packet copy from iptables.

For DNS purposes, what we really want is not to specify a limit at all; we
always want the whole DNS packet.  iptables supports that (by omitting
--nflog-size) and this PR updates Felix to omit --nflog-size for the DNS case.
As far as I'm aware there is no downside from not specifying a size, and I've
checked that the kernel copies the whole packet in that case.