Validate WEP before converting Pod to WEP
This changeset checks if a Pod is a valid WEP or not before converting it
to a WEP. Endpoints cache in the query server only reports WEP counts
back to the Manager UI. We should ignore other Pods that aren't
convertible to WEPs. This change also demotes logs to debug level to
avoid unnecessary noises when it fails to convert.