Establish VXLAN overlay between federated clusters (#6442)
This commit extends the node-to-node VXLAN mesh to include nodes from
remote clusters. This is done by extending the remote syncer to include
Node, Pool, and Block updates. Remote Pool and Block updates are keyed
by RemoteClusterResourceKey rather than their native key. This is because
their native keys have a guarantee that CIDRs within the type are disjoint,
which can't be upheld when remote clusters are connected.

Overlapping IP pools are not supported, and the implementation prevents
connection and disconnection of overlapping IP pools from disrupting 
cluster connectivity. The L3RouteResolver now indexes CIDRs in a 
cluster-aware fashion. See L3RouteResolver and VXLANManager for more 
details on conflict resolution.