Ignore old DNS info when epoch changes across a Felix restart
The DNSCacheEpoch setting is designed to be a knob that we can use to force
Felix to discard previously learned DNS mapping info.  It works if someone only
changes the DNSCacheEpoch field in their FelixConfiguration - because we have
special handling for that case to avoid a Felix restart.  But if someone
changes other FelixConfiguration fields at the same time, and those other
changes require a Felix restart (as most fields do), the DNSCacheEpoch change
is missed and the restarted Felix continues using the same DNS info as before
the restart.

This change fixes that by encoding the epoch into the mappings cache file.