Make Linseed fill in v1.RuntimeReport's GeneratedTime field
... to `time.Now()`, when writing to ES.  This delivers a major benefit of
GeneratedTime becoming monotonic across all runtime reports for a given managed
cluster - which in turn means that it becomes easy for a client to query "all
reports that were added to ES since my last query".  (Not quite trivial,
because it can get reports with exactly the last timestamp again - but those
can be easily filtered out based on the ID values.)

Note, for now we keep GeneratedTime as a pointer - even though all future
runtime reports _will_ have a GeneratedTime field - because we still want to
support Linseed reading back pre-existing reports from ES that do not have a
GeneratedTime.  There might be alternative solutions for this that don't need
GeneratedTime to remain as a pointer - especially when we know that Sasha can
only query for legacy reports back to (now - 25 minutes) - but we can work on
that, when motivated, separately from the current change.