Return events count in ServiceGraph response
This changeset reduces SG events overlay response from a full list of
events into a count. Rather than a complete list of events with details,
Manager UI only needs a flag associated with SG nodes to render
exclamation marks. The events table at the bottom of SG will send a
separate request to the backend so the SG response can be optimized to
return a count only.