Prefetch raw log data to warm up es-proxy cache
This changeset implements Service Graph prefetch in es-proxy. When
Tigera manager makes a SG request to the backend, the raw logs will be
cached in es-proxy SG cache. Subsequent requests with the same time
range and cluster name will be faster as the cache is warm. Inspired by
this behavior, SG cache will make the same request to warm up the cache
when it is initialized. In MCM or CC environment, it will also warm up
SG caches for all managed clusters.