Use upstream hcsshim instead of fork
Use github.com/Microsoft/hcsshim v0.11.4 instead of the fork github.com/projectcalico/hcsshim v0.8.9-calico.

The fork was necessary at the time of v0.8.9 in order to access the endpoint's containers, but it is no longer needed since now that information is accessible via the SharedContainers field in the HNSEndpoint struct in the upstream hcsshim package.