Add long (forward) jump capability to BPF assembler.
Write a "trampoline" block every ~30k instructions that
contains a jump for each active label.  Instructions that
jumpt to that label are fixes up eagerly to jump to the
trampoline.  Then the trampoline jumps are left for future
fixup, either jumping to the next trampoline or to the
eventual destination.