Fix collector preDNAT like connection test. (#6893)
I'm not sure what the original intention of the
test was but the source of the flake is that the
conntrack entry sent to the collector had a different
pre-DNAT dest IP to the packet info that was sent.

Depending on which arrived first, we got different
results.  Giving them the same IP in that slot makes
the output consistent.

Adjust the test so that both orders are tried
deterministically so we don't get a flake here
again.