Fix TLS metrics test.
Broken by refactoring of metrics test machinery in
ab5da6f17585522026125e9976c292bdb183c265.

Before that change it wasn't an error to get a
non-existent metric, it would just return "". Turns out
that tightening up broke a test that relied on the
old behaviour; the starts reporter seems to serve empty
metrics at start of day but the test was only verifying
connectivity so we can use GetRawMetrics instead.