app-policy: change statscache to use a callback pattern (#6937)
statscache used a channel-reading-a-channel pattern which was the root cause of flakes for dpstats in dikastes.

This PR changes the way statscache ingests and admits aggregated tuples-values pairs via direct calls and flush callbacks. sync client will now get callbacks from each statscache flush interval, keeping the old behaviour where we drop statistics when there isn't a connection to policy sync. I'm guessing the reason for this is that the stats are expected in a fix cadence.

Also added improvements to testing, updated tests where necessary.